apps/search-service/pyproject.toml
[tool.poetry]
name = "finrisk-search-service"
version = "1.0.0"
description = "FinRisk AI Copilot - Search Service with OpenSearch and hybrid vector search"
authors = ["Manikandan Bala <manikandan@finrisk.ai>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.0"

# OpenSearch and Elasticsearch
opensearch-py = "^2.4.2"
elasticsearch = "^8.11.1"

# Vector embeddings and search
sentence-transformers = "^2.2.2"
transformers = "^4.36.1"
torch = "^2.1.1"
numpy = "^1.25.2"
faiss-cpu = "^1.7.4"

# Text processing
spacy = "^3.7.2"
nltk = "^3.8.1"
textblob = "^0.17.1"
fuzzywuzzy = "^0.18.0"
python-levenshtein = "^0.23.0"

# Database
sqlalchemy = "^2.0.23"
asyncpg = "^0.29.0"
alembic = "^1.13.1"

# Cache and messaging
redis = "^5.0.1"
aiokafka = "^0.10.0"

# Utilities
pydantic-settings = "^2.1.0"
structlog = "^23.2.0"
httpx = "^0.25.2"
tenacity = "^8.2.3"
click = "^8.1.7"

# Monitoring
prometheus-client = "^0.19.0"
opentelemetry-api = "^1.21.0"
opentelemetry-sdk = "^1.21.0"
opentelemetry-instrumentation-fastapi = "^0.42b0"

# Security
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

# File processing
python-multipart = "^0.0.6"
PyPDF2 = "^3.0.1"
python-docx = "^1.1.0"
openpyxl = "^3.1.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"
pre-commit = "^3.6.0"
jupyter = "^1.0.0"

[tool.poetry.group.test.dependencies]
testcontainers = "^3.7.1"
factory-boy = "^3.3.0"
faker = "^20.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["app"]

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
asyncio_mode = "auto"
